+++ .psscripts/install-dotnet.ps1
@@ -2,9 +2,26 @@
 # Install .NET Core SDK
 # ----------------------------------------------
 
+param
+(
+    [switch] $ForceUbuntuInstall
+)
+
 $ErrorActionPreference = "Stop"
 
 Import-module "$PSScriptRoot\build-functions.ps1" -Force
+
+if ($ForceUbuntuInstall.IsPresent)
+{
+    $desiredSdk    = Get-DesiredSdk
+    $versionParts  = $desiredSdk.Split(".")
+    $majorMinorVer = $versionParts[0] + "." + $versionParts[1]
+    $ubuntuVersion = Get-UbuntuVersion
+
+    Write-Host "Ubuntu version: $ubuntuVersion"
+    Install-NetCoreSdkForUbuntu $ubuntuVersion $majorMinorVer
+    return
+}
 
 # Rename the global.json before making the dotnet --version call
 # This will prevent AppVeyor to fail because it might not find
@@ -27,10 +44,11 @@ if ($desiredSdk -eq $currentSdk)
 }
 
 Write-Host "The current .NET SDK ($currentSdk) doesn't match the project's desired .NET SDK ($desiredSdk)." -ForegroundColor Yellow
+
 Write-Host "Attempting to download and install the correct .NET SDK..."
 
 $sdkZipPath = Get-NetCoreSdkFromWeb $desiredSdk
-Install-NetCoreSdk $sdkZipPath
+Install-NetCoreSdkFromArchive $sdkZipPath
 
 Write-Host ".NET SDK installation complete." -ForegroundColor Green
 dotnet-version