+++ samples/SampleApp/SampleApp/HtmlViews.fs
@@ -6,19 +6,19 @@ open SampleApp.Models
 let layout (content: XmlNode list) =
     html [] [
         head [] [
-            title []  [ encodedText "Giraffe" ]
+            title []  [ str "Giraffe" ]
         ]
         body [] content
     ]
 
 let partial () =
-    p [] [ encodedText "Some partial text." ]
+    p [] [ str "Some partial text." ]
 
 let personView (model : Person) =
     [
-        div [_class "container"] [
-                h3 [_title "Some title attribute"] [ sprintf "Hello, %s" model.Name |> encodedText ]
-                a [_href "https://github.com/giraffe-fsharp/Giraffe"] [encodedText "Github"]
+        div [ _class "container" ] [
+                h3 [_title "Some title attribute"] [ sprintf "Hello, %s" model.Name |> str ]
+                a [ _href "https://github.com/giraffe-fsharp/Giraffe" ] [ str "Github" ]
             ]
-        div [] [partial()]
+        div [] [ partial() ]
     ] |> layout